10 Interfaces and Classes have been serated in file FormIClass.tsx
